Refactored TokenRequestListener interface to use Request in place of TokenRequest
