+++ sdk/src/main/java/co/omise/android/api/TokenRequestListener.java
@@ -1,5 +1,7 @@
-package co.omise.android;
+package co.omise.android.api;
 
+import co.omise.android.api.Request;
+import co.omise.android.api.TokenRequest;
 import co.omise.android.models.Token;
 
 /**
@@ -10,17 +12,17 @@ public interface TokenRequestListener {
      * Invoked when {@link TokenRequest} succeeds.
      *
      * @param request The original request.
-     * @param token The token result.
+     * @param token   The token result.
      */
-    void onTokenRequestSucceed(TokenRequest request, Token token);
+    void onTokenRequestSucceed(Request request, Token token);
 
     /**
      * Invoked when {@link TokenRequest} fails.
-     *
+     * <p>
      * Possible errors includes {@link org.json.JSONException} and general {@link java.io.IOException}
      *
-     * @param request The original request.
+     * @param request   The original request.
      * @param throwable The error.
      */
-    void onTokenRequestFailed(TokenRequest request, Throwable throwable);
+    void onTokenRequestFailed(Request request, Throwable throwable);
 }
