+++ sdk/src/main/java/co/omise/android/ui/CreditCardActivity.java
@@ -20,8 +20,8 @@ import co.omise.android.CardIO;
 import co.omise.android.CardNumber;
 import co.omise.android.api.Client;
 import co.omise.android.R;
-import co.omise.android.TokenRequest;
-import co.omise.android.TokenRequestListener;
+import co.omise.android.api.TokenRequest;
+import co.omise.android.api.TokenRequestListener;
 import co.omise.android.models.APIError;
 import co.omise.android.models.CardBrand;
 import co.omise.android.models.Token;
